@@ -1,39 +1,35 @@
 {
     "private": true,
     "scripts": {
-        "development": "mix",
-        "dev": "npm run development",
-        "watch": "mix watch",
-        "serve": "npm run watch",
-        "watch-poll": "mix watch -- --watch-options-poll=1000",
-        "hot": "mix watch --hot",
-        "prod": "npm run production",
-        "build": "npm run production",
-        "production": "mix --production"
+        "dev": "vite",
+        "build": "vite build"
     },
+    "type": "module",
     "devDependencies": {
-        "axios": "^0.21.1",
+        "@vitejs/plugin-vue": "^5.0.4",
+        "axios": "^1.6.7",
         "bootstrap": "^4.5.3",
         "bootstrap-datepicker": "^1.8.0",
-        "bootstrap-vue": "^2.18.0",
+        "bootstrap-vue-next": "^0.16.4",
         "clean-webpack-plugin": "^3.0.0",
         "cross-env": "^5.1",
         "jquery": "^3.5.1",
+        "laravel-vite-plugin": "^1.0.1",
         "lodash": "^4.17.21",
         "moment": "^2.29.1",
         "popper.js": "^1.12",
-        "resolve-url-loader": "^3.1.2",
+        "resolve-url-loader": "^5.0.0",
         "sass": "^1.27.0",
         "sass-loader": "^8.0.2",
-        "vue": "^2.6.12",
-        "vue-loader": "^15.11.1",
+        "vite": "^5.1.3",
+        "vue": "^3.4.19",
+        "vue-loader": "^17.4.2",
         "vue-router": "^3.4.7",
         "vue-template-compiler": "^2.6.12",
         "vuejs-datepicker": "^1.2.0",
         "vuex": "^3.5.1"
     },
     "dependencies": {
-        "autotrack": "^2.4.1",
-        "laravel-mix": "^6.0.43"
+        "autotrack": "^2.4.1"
     }
 }
