@@ -1,18 +1,9 @@
 
-window._ = require('lodash');
-window.Popper = require('popper.js').default;
+import _ from 'lodash';
+window._ = _;
 
-/**
- * We'll load jQuery and the Bootstrap jQuery plugin which provides support
- * for JavaScript based Bootstrap features such as modals and tabs. This
- * code may be modified to fit the specific needs of your application.
- */
-
-try {
-    // window.$ = window.jQuery = require('jquery');
-
-    require('bootstrap');
-} catch (e) {}
+import Popper from 'popper.js';
+window.Popper = Popper;
 
 /**
  * We'll load the axios HTTP library which allows us to easily issue requests
@@ -20,7 +11,8 @@ try {
  * CSRF token as a header based on the value of the "XSRF" token cookie.
  */
 
-window.axios = require('axios');
+import axios from 'axios';
+window.axios = axios;
 
 window.axios.defaults.headers.common['X-Requested-With'] = 'XMLHttpRequest';
 
