@@ -142,7 +142,7 @@
     import { mapGetters, mapActions, mapMutations } from 'vuex'
     import _ from 'lodash'
     import CurationNotifications from './ExistingCurationNotification.vue'
-    import DateField from '../../DateField'
+    import DateField from '../../DateField.vue'
     import curationFormMixin from '../../../mixins/curation_form_mixin'
     import ValidationError from '../../ValidationError.vue'
     import CurationStatusHistory from '../StatusHistory.vue'
